Rename ResolverMapWrapper -> ResolverMiddleware
This hopefully removes the confusion around the word wrappers for both
resolver maps and resolvers. Middlewares are a chain, or sequence, of
operations that take a copy of the previous resolver map and return a
new resolver map

Wrappers are exclusively used for resolvers now. A resolver wrapper
should take a resolver and wrap it in some way, ensuring that the
original gets called but having a way of wrapping the inputs, or
augmenting/controlling the output, of the original resolver