Move resolverMap into createGraphQLHandler options (second argument)
This API results in a cleaner usage. The resolverMap as a first arg is not always required for createGraphQLHandler like it is for pack. Being optional allows it to be specified only when it's needed in the options hash and is one less thing for people to have worry about or fill with an empty object to ignore.