Remove @graphql-mocks/paper
Turns out this middleware wasn't really needed since GraphQL Paper
works out of the box with the default GraphQL resolvers and the
naming of @graphql-mocks/paper and GraphQL Paper would probably
just add confusing. If needed the defaultResolvers for field and
type resolvers can be imported and added using the embed
middleware.