Add variable list return counts and allow for global generator options
TLDR: Only scalar fields should have data generated, but lists need
an array which is the quantity to have data generated for...

...in the case that a list is returned of objects, interfaces or unions,
the actual data itself can be handled by later resolvers but the number
that should be resolved should be determined at the current resolver.
The faker resolver will therefore, using the min/max settings, find
the return quantity and return an array of that amount but with
"undefined". When the field for the object, union or interface resolves
the fields for each "undefined" the field will look at its return type
and if its a scalar generate a value, otherwise the process repeats.

Generator options have been added at the setting for the middleware
allowing for configurability across all resolvers, or at the field level
for the specific resolver.