Shuffle around dummy-react-apps and browser acceptance tests
Setup a mock handler in network-msw that can be used for all browser-based
acceptance tests. This allows the browser acceptance tests to be shared
between network-msw and network-pretender. If this trend continues it
would be useful to pull the browser acceptance tests boilerplate out of
network-msw and abstract the pieces around the react-dummy-app a bit more
so there aren't so many floating react-dummy-apps floating around