Remove embedPackOptionsWrapper from `pack`
By including embedPackOptionsWrapper in `pack` by default with the
`embed` function, it ends up wrapping every single resolver which
can be slow if there are a lot. This case can happen if a resolver
is applied by a highlighter applied to the entire schema. In cases
where the schema is large this ends up adding a ton of additional
resolver wrapping.