Include embedPackOptionsWrapper in the layer middleware
Only include embedPackOptionsWrapper when there is at least one
wrapper included in the layer middleware arguments. If there are
no wrappers then the resolver is called with the original context
from the query. If there is at least one wrapper and it is written
in a way that it drops context, then the "managed context"
guarantee can only be guaranteed with the embedPackOptionsWrapper
which wraps the resolver ahead of any other wrappers to include
`pack` with pack options on the resolver context object.